accoutnId_auditing_lambda/requirements.txt
mysql.connector>==2.2.9
numpy>==1.21.3
boto3>==1.20.8
requests>==2.22.0
fastparquet == 0.7.2
pandas>=1.4.4
s3fs
